refactor(api): split webhook-config.service.ts into services/webhook-config/ (api-13d2)
