+++ apps/api/src/__tests__/routes/webhook-configs/restore.test.ts
@@ -13,7 +13,7 @@ import type { ApiErrorResponse } from "@pluralscape/types";
 
 // ── Mocks ────────────────────────────────────────────────────────
 
-vi.mock("../../../services/webhook-config.service.js", () => ({
+vi.mock("../../../services/webhook-config/lifecycle.js", () => ({
   restoreWebhookConfig: vi.fn(),
 }));
 vi.mock("../../../lib/audit-writer.js", () => mockAuditWriterFactory());
@@ -23,7 +23,7 @@ vi.mock("../../../middleware/auth.js", () => mockAuthFactory());
 
 // ── Imports after mocks ──────────────────────────────────────────
 
-const { restoreWebhookConfig } = await import("../../../services/webhook-config.service.js");
+const { restoreWebhookConfig } = await import("../../../services/webhook-config/lifecycle.js");
 const { systemRoutes } = await import("../../../routes/systems/index.js");
 
 // ── Helpers ──────────────────────────────────────────────────────
