+++ apps/api/src/routes/webhook-configs/archive.ts
@@ -6,7 +6,7 @@ import { createAuditWriter } from "../../lib/audit-writer.js";
 import { getDb } from "../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { archiveWebhookConfig } from "../../services/webhook-config.service.js";
+import { archiveWebhookConfig } from "../../services/webhook-config/lifecycle.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
