+++ apps/api/src/routes/webhook-configs/restore.ts
@@ -6,7 +6,7 @@ import { getDb } from "../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { restoreWebhookConfig } from "../../services/webhook-config.service.js";
+import { restoreWebhookConfig } from "../../services/webhook-config/lifecycle.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
