+++ apps/api/src/routes/webhook-configs/test.ts
@@ -5,7 +5,7 @@ import { getDb } from "../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { testWebhookConfig } from "../../services/webhook-config.service.js";
+import { testWebhookConfig } from "../../services/webhook-config/test.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
