+++ apps/api/src/routes/webhook-configs/update.ts
@@ -7,7 +7,7 @@ import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { parseJsonBody } from "../../lib/parse-json-body.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { updateWebhookConfig } from "../../services/webhook-config.service.js";
+import { updateWebhookConfig } from "../../services/webhook-config/update.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
